Move `ResourceUtils` to `org.junit.platform.commons.util`
